# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Talent Graph API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates with optional filters
      parameters:
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
        - { name: source, in: query, required: false, schema: { type: string } }
        - { name: skill, in: query, required: false, schema: { type: string } }
        - { name: location, in: query, required: false, schema: { type: string } }
        - { name: minComp, in: query, required: false, schema: { type: number } }
        - { name: maxComp, in: query, required: false, schema: { type: number } }
        - { name: sort, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CandidateListItem" }
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CandidateInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Candidate" }
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get full candidate profile
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Candidate profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Candidate" }
        "404": { description: Not found }
    put:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update candidate
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CandidateUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Candidate" }
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete candidate
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
  /candidates/{id}/trust-graph:
    get:
      operationId: getCandidateTrustGraph
      tags: [trust]
      summary: Get trust edges around a candidate
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Trust subgraph
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrustGraph" }
  /candidates/{id}/dossier/{roleId}:
    get:
      operationId: getCandidateDossier
      tags: [candidates]
      summary: Generate dossier data for a candidate-role pair
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: roleId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Dossier data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dossier" }
        "404": { description: Not found }
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List roles with optional filters
      parameters:
        - { name: status, in: query, required: false, schema: { type: string } }
        - { name: department, in: query, required: false, schema: { type: string } }
        - { name: level, in: query, required: false, schema: { type: string } }
        - { name: clientId, in: query, required: false, schema: { type: string } }
        - { name: published, in: query, required: false, schema: { type: boolean } }
      responses:
        "200":
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RoleListItem" }
    post:
      operationId: createRole
      tags: [roles]
      summary: Create a role
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Role" }
...
  /placements:
    get:
      operationId: listPlacements
      tags: [placements]
      summary: List placements
      responses:
        "200":
          description: List of placements
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Placement" }
    post:
      operationId: createPlacement
      tags: [placements]
      summary: Create a placement (calculates fee, updates statuses)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlacementInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Placement" }
...
  /client-portal/dossier/{candidateId}/{roleId}:
    get:
      operationId: getClientDossier
      tags: [clientPortal]
      summary: Read-only dossier for a candidate shortlisted on a client role
      parameters:
        - { name: candidateId, in: path, required: true, schema: { type: string } }
        - { name: roleId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Dossier data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dossier" }
...
components:
  schemas:
...
    ExperienceInput:
      type: object
      required: [company, title, startDate]
      properties:
        company: { type: string }
        title: { type: string }
        startDate: { type: string }
        endDate: { type: ["string", "null"] }
        promoted: { type: boolean }
        description: { type: ["string", "null"] }
        order: { type: integer }
    Skill:
      type: object
      required: [id, name, category, proficiency]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string, enum: [PRODUCT, TECHNICAL, LEADERSHIP, DOMAIN] }
        proficiency: { type: integer }
    SkillInput:
      type: object
      required: [name, category, proficiency]
      properties:
        name: { type: string }
        category: { type: string, enum: [PRODUCT, TECHNICAL, LEADERSHIP, DOMAIN] }
        proficiency: { type: integer }
    CandidateListItem:
      type: object
      required: [id, name, status, source]
      properties:
        id: { type: string }
        name: { type: string }
        avatarUrl: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        currentTitle: { type: ["string", "null"] }
        currentCompany: { type: ["string", "null"] }
        topSkills:
          type: array
          items: { $ref: "#/components/schemas/Skill" }
        status: { type: string, enum: [AVAILABLE, PASSIVE, PLACED, NOT_LOOKING] }
        source: { type: string, enum: [DIRECT_SEARCH, REFERRAL, COMMUNITY, INBOUND] }
        compositeScore: { type: ["number", "null"] }
        careerVelocity: { type: ["number", "null"] }
        compensationExpected: { type: ["number", "null"] }
        updatedAt: { type: string }
    Candidate:
      type: object
      required: [id, name, status, source, profileCompleteness, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        linkedinUrl: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        experiences:
          type: array
          items: { $ref: "#/components/schemas/Experience" }
        skills:
          type: array
          items: { $ref: "#/components/schemas/Skill" }
        arrInfluenced: { type: ["number", "null"] }
        revenueImpact: { type: ["string", "null"] }
        productsLaunched: { type: ["integer", "null"] }
        usersImpacted: { type: ["integer", "null"] }
        impactSummary: { type: ["string", "null"] }
        teamSize: { type: ["integer", "null"] }
        hiringResponsibility: { type: boolean }
        budgetOwnership: { type: ["string", "null"] }
        executiveExposure: { type: ["string", "null"] }
        peopleManaged: { type: ["integer", "null"] }
        directReportsCurrently: { type: ["integer", "null"] }
        compensationCurrent: { type: ["number", "null"] }
        compensationExpected: { type: ["number", "null"] }
        remotePreference: { type: string, enum: [REMOTE, HYBRID, ONSITE, FLEXIBLE] }
        relocationOpen: { type: boolean }
        industryInterests:
          type: array
          items: { type: string }
        noticePeriod: { type: ["integer", "null"] }
        status: { type: string, enum: [AVAILABLE, PASSIVE, PLACED, NOT_LOOKING] }
        source: { type: string, enum: [DIRECT_SEARCH, REFERRAL, COMMUNITY, INBOUND] }
        careerVelocity: { type: ["number", "null"] }
        profileCompleteness: { type: integer }
        recruiterNotes: { type: ["string", "null"] }
        scores:
          type: array
          items: { $ref: "#/components/schemas/ScoreHistory" }
        createdAt: { type: string }
        updatedAt: { type: string }
    CandidateInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        linkedinUrl: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        experiences:
          type: array
          items: { $ref: "#/components/schemas/ExperienceInput" }
        skills:
          type: array
          items: { $ref: "#/components/schemas/SkillInput" }
        arrInfluenced: { type: ["number", "null"] }
        revenueImpact: { type: ["string", "null"] }
        productsLaunched: { type: ["integer", "null"] }
        usersImpacted: { type: ["integer", "null"] }
        impactSummary: { type: ["string", "null"] }
        teamSize: { type: ["integer", "null"] }
        hiringResponsibility: { type: boolean }
        budgetOwnership: { type: ["string", "null"] }
        executiveExposure: { type: ["string", "null"] }
        peopleManaged: { type: ["integer", "null"] }
        directReportsCurrently: { type: ["integer", "null"] }
        compensationCurrent: { type: ["number", "null"] }
        compensationExpected: { type: ["number", "null"] }
        remotePreference: { type: string, enum: [REMOTE, HYBRID, ONSITE, FLEXIBLE] }
        relocationOpen: { type: boolean }
        industryInterests:
          type: array
          items: { type: string }
        noticePeriod: { type: ["integer", "null"] }
        status: { type: string, enum: [AVAILABLE, PASSIVE, PLACED, NOT_LOOKING] }
        source: { type: string, enum: [DIRECT_SEARCH, REFERRAL, COMMUNITY, INBOUND] }
        recruiterNotes: { type: ["string", "null"] }
    CandidateUpdate:
      allOf:
        - $ref: "#/components/schemas/CandidateInput"
...
    RoleListItem:
      type: object
      required: [id, title, level, department, status, clientName]
      properties:
        id: { type: string }
        title: { type: string }
        clientId: { type: string }
        clientName: { type: string }
        level: { type: string }
        department: { type: string }
        compensationMin: { type: number }
        compensationMax: { type: number }
        feePercentage: { type: number }
        location: { type: ["string", "null"] }
        status: { type: string }
        daysOpen: { type: integer }
        pipelineCount: { type: integer }
        publishedToNetwork: { type: boolean }
    Role:
      type: object
      required: [id, title, level, department, status, clientId, clientName, createdAt]
      properties:
        id: { type: string }
        clientId: { type: string }
        clientName: { type: string }
        title: { type: string }
        level: { type: string, enum: [IC, MANAGER, SENIOR_MANAGER, DIRECTOR, VP, CXO] }
        department: { type: string, enum: [PRODUCT, ENGINEERING, DESIGN, DATA, GENERAL_MANAGEMENT] }
        compensationMin: { type: number }
        compensationMax: { type: number }
        feePercentage: { type: number }
        location: { type: ["string", "null"] }
        remotePolicy: { type: string, enum: [REMOTE, HYBRID, ONSITE, FLEXIBLE] }
        mustHaveSkills:
          type: array
          items: { type: string }
        niceToHaveSkills:
          type: array
          items: { type: string }
        experienceMin: { type: integer }
        experienceMax: { type: integer }
        industryPreference:
          type: array
          items: { type: string }
        teamContext: { type: ["string", "null"] }
        successCriteria: { type: ["string", "null"] }
        antiPatterns:
          type: array
          items: { type: string }
        impactThreshold: { type: ["string", "null"] }
        leadershipSignal: { type: ["string", "null"] }
        cultureFit: { type: ["string", "null"] }
        status: { type: string, enum: [DISCOVERY, ACTIVE, SHORTLISTING, INTERVIEWING, OFFER, PLACED, CANCELLED] }
        publishedToNetwork: { type: boolean }
        maxShortlist: { type: integer }
        daysOpen: { type: integer }
        pipelineCount: { type: integer }
        stageCounts:
          type: array
          items: { $ref: "#/components/schemas/StageCount" }
        estimatedFee: { type: number }
        createdAt: { type: string }
    RoleInput:
      type: object
      required: [clientId, title, level, department, compensationMin, compensationMax, feePercentage, experienceMin, experienceMax]
      properties:
        clientId: { type: string }
        title: { type: string }
        level: { type: string, enum: [IC, MANAGER, SENIOR_MANAGER, DIRECTOR, VP, CXO] }
        department: { type: string, enum: [PRODUCT, ENGINEERING, DESIGN, DATA, GENERAL_MANAGEMENT] }
        compensationMin: { type: number }
        compensationMax: { type: number }
        feePercentage: { type: number }
        location: { type: ["string", "null"] }
        remotePolicy: { type: string, enum: [REMOTE, HYBRID, ONSITE, FLEXIBLE] }
        mustHaveSkills:
          type: array
          items: { type: string }
        niceToHaveSkills:
          type: array
          items: { type: string }
        experienceMin: { type: integer }
        experienceMax: { type: integer }
        industryPreference:
          type: array
          items: { type: string }
        teamContext: { type: ["string", "null"] }
        successCriteria: { type: ["string", "null"] }
        antiPatterns:
          type: array
          items: { type: string }
        impactThreshold: { type: ["string", "null"] }
        leadershipSignal: { type: ["string", "null"] }
        cultureFit: { type: ["string", "null"] }
        status: { type: string, enum: [DISCOVERY, ACTIVE, SHORTLISTING, INTERVIEWING, OFFER, PLACED, CANCELLED] }
        publishedToNetwork: { type: boolean }
        maxShortlist: { type: integer }
...
    StageCount:
      type: object
      required: [stage, count]
      properties:
        stage: { type: string }
        count: { type: integer }
...
    TrustGraph:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items: { $ref: "#/components/schemas/TrustNode" }
        edges:
          type: array
          items: { $ref: "#/components/schemas/TrustEdge" }
...
    Placement:
      type: object
      required: [id, roleId, roleTitle, candidateId, candidateName, clientId, clientName, actualCompensation, feeAmount, status, createdAt]
      properties:
        id: { type: string }
        roleId: { type: string }
        roleTitle: { type: string }
        candidateId: { type: string }
        candidateName: { type: string }
        clientId: { type: string }
        clientName: { type: string }
        actualCompensation: { type: number }
        feePercentage: { type: number }
        feeAmount: { type: number }
        estimatedFee: { type: ["number", "null"] }
        offerDate: { type: ["string", "null"] }
        joinDate: { type: ["string", "null"] }
        status: { type: string, enum: [CONFIRMED, JOINED, PROBATION_PASSED, REPLACEMENT_TRIGGERED] }
        guaranteeExpiry: { type: ["string", "null"] }
        createdAt: { type: string }
    PlacementInput:
      type: object
      required: [roleId, candidateId, actualCompensation]
      properties:
        roleId: { type: string }
        candidateId: { type: string }
        actualCompensation: { type: number }
        offerDate: { type: ["string", "null"] }
        joinDate: { type: ["string", "null"] }
...
    Dossier:
      type: object
      required: [candidate, role, score, professionalSummary]
      properties:
        candidate: { $ref: "#/components/schemas/Candidate" }
        role: { $ref: "#/components/schemas/Role" }
        score: { $ref: "#/components/schemas/CandidateScore" }
        professionalSummary: { type: string }
        trustIndicators:
          type: array
          items: { $ref: "#/components/schemas/TrustEdge" }
        endorsers:
          type: array
          items: { $ref: "#/components/schemas/TrustNode" }
...